pyproject.toml
[project]
name = "leno-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.5.3",
    "django>=5.2.7",
    "django-cors-headers>=4.9.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "httpx>=0.28.1",
    "ipython>=9.6.0",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "django-debug-toolbar>=6.1.0",
    "ipdb>=0.13.13",
]
